requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
google-generativeai==0.8.5
python-dotenv>=1.0.0
python-multipart==0.0.6
pydantic-settings>=2.0.0
pgvector==0.2.1
sqlalchemy[postgresql]>=2.0.0